refactor: restructure startup for cleaner main.rs
refactor(service): streamline config load error handling

refactor: initialize tracing subscriber in main

refactor(main): use canonical SUCCESS AND FAILURE exit
