+++ service/src/lib.rs
@@ -1,0 +1,9 @@
+// Copyright 2023-, GraphOps and Semiotic Labs.
+// SPDX-License-Identifier: Apache-2.0
+
+mod cli;
+mod config;
+mod database;
+mod error;
+mod routes;
+pub mod service;
